requirements.txt
Pillow>=10.0,<11
matplotlib>=3.7,<3.8; python_version < "3.9"
matplotlib>=3.8,<4; python_version >= "3.9"
nbformat>=5.9,<6
nbclient>=0.9,<0.11
nbconvert>=7.16,<8
notebook>=6.4,<6.5; python_version < "3.9"
notebook>=7.0,<8; python_version >= "3.9"
jupyter-client>=7.4,<8; python_version < "3.9"
pywinpty>=1.1,<2; platform_system == "Windows" and python_version < "3.9"
ipykernel>=6.29,<7
